requirements.txt
Flask==3.1.0
flask-cors==5.0.1
pymongo[srv]==3.12.0
pandas==2.2.3
numpy==2.2.5
openpyxl==3.1.5
python-dotenv==1.1.0
gunicorn==23.0.0
APScheduler==3.11.0
openai==1.60.2
requests==2.32.3